\name{AICc2LL}
\alias{AICc2LL}
\title{
Convert an AICc value to a raw log likelihood
}
\description{
Convert an AICc value generated by fitting a theoretical abundance distribution to a raw log likelihood.
}
\usage{
AICc2LL(n,AICc,k=1)
}
\arguments{
  \item{n}{
The list of counts.
}
  \item{AICc}{
The AICc to convert.
}
  \item{k}{
The number of parameters in the model (either 1 or 2).
}
}
\details{
This function takes output from any of the distribution fitting routines included in the \emph{richness} package and converts it to the original log likelihood value.
}
\value{
 {The converted log likelihood.}
}
\author{
John Alroy
}
\examples{
# generate a random list of non-zero counts from the scaled odds distribution
n <- rodds(100)
n <- n[n > 0]

# fit the scaled odds distribution and save the AICc
a <- sodds(n)$AICc

# convert to a log likelihood
AICc2LL(n,a,2)

}
