\name{cegsB}
\alias{cegsB}
\title{
The compound exponential-geometric series fit by a Bayesian calculation
}
\description{
Uses a Bayesian calculation to fit a species abundance distribution based on a count vector to the compound exponential-geometric series.
}
\usage{
cegsB(n)
}
\arguments{
  \item{n}{
A vector of integer counts representing the number of individuals per species.
}
}
\details{
This function fits the compound exponential-geometric series (CEGS) distribution, returning the corrected Akaike information criterion (AICc) fit statistic and the rank and species abundance distributions in addition to estimated richness and the distribution's two governing parameters.

CEGS assumes each count is generated by a random draw from the geometric series (GS), but with the governing parameter \emph{p} of the GS being itself a function of a random exponential variate \emph{E} = -log(\emph{U}), where \emph{U} is uniformly distributed between 0 and 1. In other words, each count is doubly random. The expression is fitted to data by varying a scale term \eqn{\lambda} and a shape term \eqn{\gamma}, which feature in the probability mass function at different points.

Further details that include the derivation of the compound distribution are given on the \link{cegsML} manual page.

The Bayesian calculation assumes that the priori distributions of \eqn{\lambda} and \eqn{\gamma} are produced by taking negative logs of fractions that are evenly spaced between 0 and 1. A grid computation is used to obtain the likelihoods of the count distributions given possible combinations of the parameters (i.e., the conditional probabilities). To do this, a list of values spaced equally from 0 to 1 is created for each parameter. The likelihoods are scaled so they sum to 1; they are multiplied across the lists of uniformly spaced values; the sum is taken; and finally, the negative log of the sum is taken to obtain the posterior estimate.

Standard ecological data tend to yield \eqn{\gamma} values that have less variance than \eqn{\lambda} values, which is unsurprising because \eqn{\gamma} is a power term. Therefore, the parameter grid spans 30 points for \eqn{\gamma} and 90 for \eqn{\lambda}. This yields limits of -log(30/31) = 0.033 to -log(1/31) = 3.434 for \eqn{\gamma}, and -log(90-/91) = 0.011 to -log(1/91) = 4.511 for \eqn{\lambda}.

An AICc is also reported for comparability with other functions. \link{cegsML} will produce lower AICcs if a likelihood approach is preferred. The likelihood calculation that yields the AICc is explained in the package details.

The function \link{rcegs} computes random integers based on the distribution.
}
\value{
  \item{richness}{The sample's estimated species richness.}
  \item{scale}{The estimated \eqn{\lambda} (scale) parameter of the CEGS distribution.}
  \item{shape}{The estimated \eqn{\gamma} (shape) parameter of the CEGS distribution.}
  \item{AICc}{The corrected Akaike information criterion value computed from the log likelihood.}
  \item{fitted.RAD}{The predicted rank abundance distribution.}
  \item{fitted.SAD}{The predicted species abundance distribution.}
}
\author{
John Alroy
}
\examples{
# CEGS distribution
n <- rcegs(100)
cegsB(n[n > 0])$richness

# Poisson log normal distribution
n <- rpois(100,exp(rnorm(100,sd=2)))
cegsB(n[n > 0])$richness

# scaled odds distribution
n <- rodds(100,1)
cegsB(n[n > 0])$richness
}
