\name{exp2e}
\alias{exp2e}
\title{
The exponential-to-\emph{e} distribution fit by maximum likelihood
}
\description{
Uses maximum likelihood to fit a species abundance distribution based on a count vector to the exponential-to-\emph{e} distribution.
}
\usage{
exp2e(n)
}
\arguments{
  \item{n}{
A vector of integer counts representing the number of individuals per species.
}
}
\details{
This function fits the exponential-to-\emph{e} (exp2e) distribution, returning the corrected Akaike information criterion (AICc) fit statistic and the rank and species abundance distributions in addition to estimated richness and the distribution's single governing parameter.

exp2e assumes each count is generated by rounding down a value drawn from an exponential distribution raised to the power of \emph{e} and with a \emph{lambda} (rate) scaling parameter (l in the code). The distribution usually generates low variation in counts.

As with the scaled odds distribution (\link{sodds}), exp2e can be matched precisely with data derived from a simple birth-death simulation. The model is fully neutral except that it assumes individual species have birth and death rates that are fixed and correlated. It is expected to be realistic whenever (1) traits determine demographic properties and (2) the environment shows little variation.

The analytical expression for the probability mass function (PMF) is: 
 
\code{
-diff(exp(-l * x^exp(-1))) / exp(-l)
}

where l = \emph{lambda} and x = the positive integer series starting with 1, which corresponds with the count classes (e.g., classes 1 and 2 are the respective counts of singletons and doubletons). The equation assumes that all counts are greater than zero.

The model is fit by optimising the AICc using the \emph{stats4} function \emph{lme}, as with this package's other functions.
The likelihood calculation that yields the AICc is explained in the package details.

The function \link{rexp2e} computes random integers based on the distribution.
}
\value{
  \item{richness}{The sample's estimated species richness.}
  \item{lambda}{The estimated governing parameter of the exponential-to-\emph{e} distribution.}
  \item{AICc}{The corrected Akaike information criterion value computed from the log likelihood.}
  \item{fitted.RAD}{The predicted rank abundance distribution.}
  \item{fitted.SAD}{The predicted species abundance distribution.}
}
\author{
John Alroy
}
\examples{
# Poisson log normal distribution
n <- rpois(100,exp(rnorm(100,sd=2)))
exp2e(n[n > 0])$richness

# scaled odds distribution
n <- rodds(100,1)
exp2e(n[n > 0])$richness

# exponential-to-e distribution
n <- rexp2e(100,1)
exp2e(n[n > 0])$richness
}
