% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{fitBS}
\alias{fitBS}
\title{Fits the broken stick distribution}
\usage{
fitBS(n)
}
\arguments{
\item{n}{Vector of species abundances}
}
\value{
A list with the following components:
\describe{
\item{distribution}{The XXX}
\item{richness}{The XXX}
\item{fit}{The parameter XXX}
}
}
\description{
Fits the broken stick distribution
}
\examples{
\dontrun{
fitBS(n)
}
}
