% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{fitDoubleGeometric}
\alias{fitDoubleGeometric}
\title{Fits the double geometric distribution}
\usage{
fitDoubleGeometric(n, trials = 1000, min = NA, max = NA)
}
\arguments{
\item{n}{Vector of species abundances}

\item{trials}{Number of interactions when fitting function.}

\item{min}{Minimum bound on total species richness}

\item{max}{Maximum bound on total species richness}
}
\value{
A list with the following components:
\describe{
\item{distribution}{The XXX}
\item{richness}{The XXX}
\item{k}{The parameter XXX}
}
}
\description{
Fits the double geometric distribution
}
\examples{
\dontrun{
fitDoubleGeometric(n)
}
}
\references{
Alroy, J. 2015. The shape of terrestrial abundance distributions. Science Advances 1(8): e1500082
}
