% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{fitGeometric}
\alias{fitGeometric}
\title{Fits the geometric series distribution}
\usage{
fitGeometric(n)
}
\arguments{
\item{n}{Vector of species abundances}
}
\value{
A list with the following components:
\describe{
\item{distribution}{The XXX}
\item{k}{The parameter XXX}
\item{fit}{The XXX}
}
}
\description{
Fits the geometric series distribution
}
\examples{
\dontrun{
fitGeometric(n)
}
}
