% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{fitLogNormal}
\alias{fitLogNormal}
\title{Fits the log normal distribution}
\usage{
fitLogNormal(n, trials = 1000, min = NA, max = NA)
}
\arguments{
\item{n}{Vector of species abundances}

\item{trials}{Number of interactions when fitting function.}

\item{min}{Minimum bound on total species richness}

\item{max}{Maximum bound on total species richness}
}
\value{
A list with the following components:
\describe{
\item{distribution}{The XXX}
\item{richness}{The XXX}
\item{sd}{The parameter XXX}
}#' @export
}
\description{
Fits the log normal distribution
}
\examples{
\dontrun{
fitLogNormal(n)
}
}
