% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{fitLogSeries}
\alias{fitLogSeries}
\title{Fits the log series distribution}
\usage{
fitLogSeries(n)
}
\arguments{
\item{n}{Vector of species abundances}
}
\value{
A list with the following components:
\describe{
\item{distribution}{The XXX}
\item{alpha}{The XXX}
\item{fit}{The parameter XXX}
}
}
\description{
Fits the log series distribution
}
\examples{
\dontrun{
fitLogSeries(n)
}
}
