% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{fitZipf}
\alias{fitZipf}
\title{Fits the Zipf distribution}
\usage{
fitZipf(n)
}
\arguments{
\item{n}{Vector of species abundances}
}
\value{
A list with the following components:
\describe{
\item{distribution}{The XXX}
\item{exponent}{The XXX}
\item{fit}{The XXX}
}
}
\description{
Fits the Zipf distribution
}
\examples{
\dontrun{
fitZipf(n)
}
}
