\name{ipower}
\alias{ipower}
\title{
The inverse power distribution fit by maximum likelihood
}
\description{
Uses maximum likelihood to fit a species abundance distribution based on a count vector to the inverse power distribution.
}
\usage{
ipower(n)
}
\arguments{
  \item{n}{
A vector of integer counts representing the number of individuals per species.
}
}
\details{
This function fits the inverse power distribution, returning the corrected Akaike information criterion (AICc) fit statistic and the rank and species abundance distributions in addition to estimated richness and the distribution's single governing parameter.

\emph{ipower} assumes each count is generated by rounding down a value drawn from an inverse power distribution. The scaling parameter is just the power to which any given count plus one is raised. The distribution usually generates high variation in counts.

As with the similar scaled odds distribution (\link{sodds}), the inverse power distribution can be matched precisely with data derived from a simple birth-death simulation. The model is fully neutral except that it assumes individual species have birth and death rates that are fixed and correlated. It is expected to be realistic whenever traits determine demographic properties.

The analytical expression for the probability mass function (PMF) is: 
 
\code{
-diff(1 / (x + 1)^b) * 2^b
}

where b = \emph{beta} and x = the positive integer series starting with 1, which corresponds with the count classes (e.g., classes 1 and 2 are the respective counts of singletons and doubletons). The equation assumes that all counts are greater than zero.

The model is fit by optimising the AICc using \emph{optimise}, as with most of this package's other functions.
The likelihood calculation that yields the AICc is explained in the package details.

The function \link{ripower} computes random integers based on the distribution.
}
\value{
  \item{richness}{The sample's estimated species richness.}
  \item{beta}{The estimated governing parameter of the inverse power exponential distribution.}
  \item{AICc}{The corrected Akaike information criterion value computed from the log likelihood.}
  \item{fitted.RAD}{The predicted rank abundance distribution.}
  \item{fitted.SAD}{The predicted species abundance distribution.}
}
\author{
John Alroy
}
\examples{
# Poisson log normal distribution
n <- rpois(100,exp(rnorm(100,sd=2)))
ipower(n[n > 0])$richness

# scaled odds distribution
n <- rodds(100,1)
ipower(n[n > 0])$richness

# inverse power exponential distribution
n <- ripower(100,0.5)
ipower(n[n > 0])$richness
}
