% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{logSeriesParams}
\alias{logSeriesParams}
\title{Helper function for fitLogSeries}
\usage{
logSeriesParams(n)
}
\arguments{
\item{n}{Vector of species abundances}
}
\value{
A vector
}
\description{
Helper function for fitLogSeries
}
\examples{
\dontrun{
logSeriesParams(n)
}
}
