\name{sodds}
\alias{sodds}
\title{
The scaled odds distribution fit by maximum likelihood
}
\description{
Uses maximum likelihood to fit a species abundance distribution based on a count vector to the scaled odds distribution.
}
\usage{
sodds(n)
}
\arguments{
  \item{n}{
A vector of integer counts representing the number of individuals per species.
}
}
\details{
This function fits the scaled odds distribution, returning the corrected Akaike information criterion (AICc) fit statistic and the rank and species abundance distributions in addition to estimated richness and the distribution's single governing parameter.

The scaled odds distribution assumes each count is generated by rounding down a value drawn from random odds that are centered with a constant \emph{mu} (m in the code). Scaled odds are in the form \emph{mu} (1 - \emph{U})/\emph{U} where \emph{U} is a random uniform variate. The distribution often generates high variation in counts.

As with the exponential-to-\emph{e} distribution (\link{exp2e}), scaled odds can be matched precisely with counts derived from a simple birth-death simulation. The model is fully neutral and assumes that demographic rates are randomly variable both among species and through time. It is expected to be realistic if the environment shows high variation and if different species are adapted better or worse to certain environmental conditions.

The analytical expression for the probability mass function (PMF) is:

\code{
(m + 1)/((m + x) * (m + x + 1))
}

where m = \emph{mu} and x = the positive integer series starting with 1, which corresponds with the count classes (e.g., classes 1 and 2 are the respective counts of singletons and doubletons). The equation assumes that all counts are greater than zero.

The model is fit by optimising the AICc using the \emph{stats4} function \emph{lme}, as with this package's other functions.
The likelihood calculation that yields the AICc is explained in the package details.

The function \link{rodds} computes random integers based on the distribution.
}
\value{
  \item{richness}{The sample's estimated species richness.}
  \item{mu}{The estimated governing parameter of the scaled odds distribution.}
  \item{AICc}{The corrected Akaike information criterion value computed from the log likelihood.}
  \item{fitted.RAD}{The predicted rank abundance distribution.}
  \item{fitted.SAD}{The predicted species abundance distribution.}
}
\author{
John Alroy
}
\examples{
# Poisson log normal distribution
n <- rpois(100,exp(rnorm(100,sd=2)))
sodds(n[n > 0])$richness

# scaled odds distribution
n <- rodds(100,1)
sodds(n[n > 0])$richness

# exponential-to-e distribution
n <- rexp2e(100,1)
sodds(n[n > 0])$richness
}
